requirements.txt
# Bird Ringing Dashboard - Requirements

# Core dependencies
dash>=2.14.0
plotly>=5.18.0
pandas>=2.1.0
polars>=0.19.0
duckdb>=0.9.0

# Data processing
numpy>=1.24.0
scikit-learn>=1.3.0

# Web server
gunicorn>=21.2.0

# Development
python-dotenv>=1.0.0

# Optional: for advanced visualizations
kaleido>=0.2.1  # For static image export

pyarrow==23.0.0  # For Parquet support in Polars